Refactored styles to always have a navbar
