+++ client/src/components/Landing/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-// import { Link } from 'react-router-dom';
 
 import * as s from './styles';
 
@@ -7,14 +6,7 @@ class Landing extends React.Component {
   render() {
     return (
       <s.LandingContainer className="App">
-        <s.NavBar>
-          <s.StyledLink to="/login">Login</s.StyledLink>
-          <s.StyledLink to="/register">Register</s.StyledLink>
-          <s.StyledLink to="/test">TEST PAGE</s.StyledLink>
-        </s.NavBar>
-        <s.Content>
-          <s.Title>Landing</s.Title>
-        </s.Content>
+        <s.Title>Landing</s.Title>
       </s.LandingContainer>
     );
   }
