+++ client/src/components/Navbar/index.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import * as s from './styles';
+
+class Navbar extends React.Component {
+  render() {
+    return (
+      <s.Navbar>
+        <s.StyledLink to="/login">Login</s.StyledLink>
+        <s.StyledLink to="/register">Register</s.StyledLink>
+        <s.StyledLink to="/test">TEST PAGE</s.StyledLink>
+      </s.Navbar>
+    );
+  }
+}
+
+export default Navbar;
