+++ client/src/components/Navbar/styles.js
@@ -1,0 +1,22 @@
+import styled from 'styled-components';
+import { Link } from 'react-router-dom';
+
+export const Navbar = styled.nav`
+  padding-right: 20px;
+  width: 100%;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  height: 50px;
+  background: red;
+`;
+
+export const StyledLink = styled(Link)`
+  color: white;
+  text-decoration: none;
+  margin-right: 10px;
+
+  &:last-child {
+    margin-right: 0;
+  }
+`;
